Updated out of date dependencies and replaced a deprecated nuget package.
- Replaced DotNetZip with SharpIOZipLib
- Fixed analyzer warning for some LINQ
- Fixed analyzer warning for Avalonia (the text skew is STILL broken)
- Fixed bug in GorgonFileExtensionCollection where Count was not implemented
- Fixed Stream files to implement the correct functions, thereby fixing bugs when reading gorgon packed files.
- Fixed issue in animation codec where texture coordinate data was not saved correctly.
-

+++ Test/Gorgon.Core.Tests/GorgonColorTests.cs
@@ -146,7 +146,7 @@ public class GorgonColorTests
 
         unchecked
         {
-            Assert.AreEqual((int)0x7fbf3fff, signedRgba);
+            Assert.AreEqual(0x7fbf3fff, signedRgba);
         }
     }
 
@@ -169,7 +169,7 @@ public class GorgonColorTests
 
         unchecked
         {
-            Assert.AreEqual((int)0x3fbf7fff, signedBgra);
+            Assert.AreEqual(0x3fbf7fff, signedBgra);
         }
     }
